% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlot.R
\name{BoxPlot}
\alias{BoxPlot}
\title{function to generate box plot cluster analysis.}
\usage{
BoxPlot(fn, df.data, color.cluster, feature, value.p, log2 = F)
}
\arguments{
\item{log2}{(bool)}
}
\description{
This function generate box plot figure filled in by group.
}
