% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.plot.colors.R
\name{get.plot.colors}
\alias{get.plot.colors}
\title{return list of n colors scaled on color palette defined}
\usage{
get.plot.colors(n = 5, color.palette = NULL, sample.color.flash = T)
}
\arguments{
\item{n}{(int) number of colors to return}

\item{color.palette}{color palette to use.}
}
\value{
list of colors
}
\description{
return list of n colors scaled on color palette defined
}
