% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapGlobalJaccardSample.R
\name{HeatmapGlobalJaccardSample}
\alias{HeatmapGlobalJaccardSample}
\title{Heatmap (n_sample * n_sample) with Jaccard distances.}
\usage{
HeatmapGlobalJaccardSample(
  jc_sample,
  nseed,
  path_to_save,
  as_pdf = T,
  as_png = F,
  file_name = "heatmap_global_jaccard_sample",
  width = 15,
  height = 13
)
}
\arguments{
\item{jc_sample}{matrix with heatmap information from pheatmap library}

\item{nseed}{(int) number of generated maps}

\item{path_to_save}{path to save figure}

\item{as_pdf}{(bool) if generate and save pdf figure}

\item{as_png}{(bool) if generated and save png figure}

\item{file_name}{name of the file to generate}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{title}{title of the figure}
}
\description{
This function create a plot with heat map information.
This is a matrix (n_sample * n_sample) with whit color for small Jaccard
distance values and red for high distances.
}
