% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JaccardScoresPlot.R
\name{JaccardScoresPlot}
\alias{JaccardScoresPlot}
\title{Plot a simple graph with specific Jaccard scores for multiple models.}
\usage{
JaccardScoresPlot(
  path_to_save,
  jaccard_index_tot,
  nb_cluster,
  score_type = "noType",
  plot_indexes = F,
  as_pdf = T,
  as_png = F,
  width = 10,
  height = 7
)
}
\arguments{
\item{path_to_save}{path to save figure}

\item{jaccard_index_tot}{data frame with all Jaccard index saved}

\item{nb_cluster}{the number of cluster}

\item{score_type}{type of score to plot}

\item{plot_indexes}{(bool) if we plot Jaccard inexes scores (specificity and
sensitivity and random index)}

\item{as_pdf}{(bool) if generate and save pdf figure}

\item{as_png}{(bool) if generated and save png figure}

\item{width}{width of the figure}

\item{height}{height of the figure}
}
\description{
This function plot a simple graph with Jaccard scores for all
models passed in argument.
This is a function to compare different model to be take a decision of which
model are more better (according this score).
}
