% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomMapCluster.R
\name{SomMapCluster}
\alias{SomMapCluster}
\title{Generate SOM map with different information about clustering}
\usage{
SomMapCluster(
  data.som,
  path.plot.save,
  groups = NULL,
  nb.cluster = 5,
  add.boundaries = T,
  add.samples = F,
  is.samples.colored = F,
  clusters = NULL,
  as.pdf = T,
  as.png = F,
  type = "property",
  property = NULL,
  title = "Cluster",
  width = 7,
  height = 7,
  main = "No title",
  color.palette = NULL,
  nb.colors = 20,
  nb.samples = NULL,
  add.pie = F,
  dt.projected = NULL,
  color.clinical.variable = NULL,
  color.bg = NULL,
  add.legend = F,
  fisher.table = NULL,
  legend = F,
  legend.txt = NULL,
  legend.fill = NULL,
  add.arrow = F,
  number.pies = 5,
  list.p.value.test = NULL,
  list.titles = NULL,
  p.value.threshold = 0.05,
  pie.labels = NA,
  list.colors = NULL
)
}
\arguments{
\item{data.som}{(.rds) RSS map with information to plot}

\item{path.plot.save}{path to save figure}

\item{groups}{(default=NULL) matrix with clustering groups information}

\item{add.boundaries}{(bool) (default=T) if we add cluster boundaries in plot}

\item{add.samples}{(bool) (default=F) if we add sample circle plot in the map}

\item{is.samples.colored}{(bool) (default=F) if we use colored sample
(according to the cluster which it belongs)}

\item{as.pdf}{(bool) if generate and save pdf figure}

\item{as.png}{(bool) if generated and save png figure}

\item{type}{(str) type of SOM plot (by default is property)}

\item{title}{title of the figure}

\item{width}{width of the figure}

\item{height}{height of the figure.}

\item{nb.colors}{(int) number of colors (equal of number of cluster)}

\item{Clusters}{(default=NULL) data frame with clustering information}
}
\description{
This function generates a figure representing a SOM map.
This function can add samples colored according to the cluster it belongs to.
It can plot also the cluster boundaries.

About add pie chart: we need to pass a data frame with pie chart information.
We need to provide some information like proportions, p value, pie name in
the data frame.
TODO: define a specific format that the data for the pie chart will have to
respect (minimal information to have + tests).
}
