% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomMapCodes.R
\name{SomMapCodes}
\alias{SomMapCodes}
\title{generate SOM map with code values in nodes.}
\usage{
SomMapCodes(
  data.som,
  path.plot.save,
  as.pdf = T,
  as.png = F,
  file.name = "Codes",
  title = "Codes",
  width = 7,
  height = 7
)
}
\arguments{
\item{data.som}{(.rds) RSS map with information to plot}

\item{path.plot.save}{path to save figure}

\item{as.pdf}{(bool) if generate and save pdf figure}

\item{as.png}{(bool) if generated and save png figure}

\item{file.name}{name of the file to generate}

\item{title}{title of the figure}

\item{width}{width of the figure}

\item{height}{height of the figure}
}
\description{
This function generate a SOM map with in each node a diagram
with factor values.
}
